Porting additional tests to MPI
