+++ test/mpi/runtests.jl
@@ -28,5 +28,8 @@ function run_driver(procs,file,sysimage)
   end
 end
 #run_driver(4,"CubedSphereDiscreteModelsTests.jl",sysimage)
-run_driver(4,"DarcyCubedSphereTests.jl",sysimage)
+#run_driver(4,"DarcyCubedSphereTests.jl",sysimage)
+run_driver(4,"LaplaceBeltramiCubedSphereTests.jl",sysimage)
+#run_driver(4,"WeakDivPerpTests.jl",sysimage)
+
 end # module
