+++ test/sequential/runtests.jl
@@ -5,6 +5,7 @@ using Test
   @time @testset "CubedSphereDiscreteModelsTests" begin include("CubedSphereDiscreteModelsTests.jl") end
   @time @testset "DarcyCubedSphereTests" begin include("DarcyCubedSphereTests.jl") end
   @time @testset "WeakDivPerpTests" begin include("WeakDivPerpTests.jl") end
+  @time @testset "WeakDivPerpTests" begin include("WeakGradTests.jl") end
   @time @testset "LaplaceBeltramiCubedSphereTests" begin include("LaplaceBeltramiCubedSphereTests.jl") end
   @time @testset "WaveEquationCubedSphereTests" begin include("WaveEquationCubedSphereTests.jl") end
   #@time @testset "Williamsom2ThetaMethodFullNewtonTests" begin include("Williamsom2ThetaMethodFullNewtonTests.jl") end
